fix: use the correct boradcastchannel name and replace the sign transactions method for legacy
