+++ src/external/services/LegacyClientMessageBroker/types/index.ts
@@ -1,3 +1,6 @@
 export type { default as IInitOptions } from './IInitOptions';
 export type { default as ILegacyClientRequestMessage } from './ILegacyClientRequestMessage';
 export type { default as ILegacyClientResponseMessage } from './ILegacyClientResponseMessage';
+export type { default as ILegacyDiscoverResult } from './ILegacyDiscoverResult';
+export type { default as INetworkConfigurationWithLegacyMethods } from './INetworkConfigurationWithLegacyMethods';
+export type { default as TLegacyResponseResults } from './TLegacyResponseResults';
